chore: set formatShareConfigs as pure function
